+++ crates/bevy_ecs/src/index/mod.rs
@@ -128,14 +128,12 @@
 //! This is great for usability, but means all mutations of indexed components will carry a small but
 //! existent overhead.
 
+mod query_by_index;
+
+pub use query_by_index::*;
+
 use crate::{
-    self as bevy_ecs,
-    component::{Component, ComponentDescriptor, ComponentId, Immutable, StorageType, Tick},
-    entity::Entity,
-    prelude::Trigger,
-    query::{QueryBuilder, QueryData, QueryFilter, QueryState, With},
-    system::{Commands, Query, Res, ResMut, SystemMeta, SystemParam},
-    world::{unsafe_world_cell::UnsafeWorldCell, OnInsert, OnReplace, World},
+    self as bevy_ecs, component::{Component, ComponentDescriptor, ComponentId, Immutable, StorageType}, entity::Entity, prelude::Trigger, query::{QueryBuilder, QueryData, QueryFilter}, system::{Commands, Query, ResMut, SystemParam}, world::{OnInsert, OnReplace, World}
 };
 use alloc::{format, vec::Vec};
 use bevy_ecs_macros::Resource;
@@ -179,153 +177,6 @@ impl WorldIndexExtension for World {
 pub trait IndexableComponent: Component<Mutability = Immutable> + Eq + Hash + Clone {}
 
 impl<C: Component<Mutability = Immutable> + Eq + Hash + Clone> IndexableComponent for C {}
-
-/// This system parameter allows querying by an [indexable component](`IndexableComponent`) value.
-///
-/// # Examples
-///
-/// ```rust
-/// # use bevy_ecs::prelude::*;
-/// # let mut world = World::new();
-/// #[derive(Component, PartialEq, Eq, Hash, Clone)]
-/// #[component(immutable)]
-/// struct Player(u8);
-///
-/// // Indexing is opt-in through `World::add_index`
-/// world.add_index::<Player>();
-/// # world.spawn(Player(0));
-/// # world.spawn(Player(0));
-/// # world.spawn(Player(1));
-/// # world.spawn(Player(1));
-/// # world.spawn(Player(1));
-/// # world.spawn(Player(2));
-/// # world.spawn(Player(2));
-/// # world.spawn(Player(2));
-/// # world.spawn(Player(2));
-/// # world.flush();
-///
-/// fn find_all_player_one_entities(mut query: QueryByIndex<Player, Entity>) {
-///     for entity in query.at(&Player(0)).iter() {
-///         println!("{entity:?} belongs to Player 1!");
-///     }
-/// #   assert_eq!(query.at(&Player(0)).iter().count(), 2);
-/// #   assert_eq!(query.at(&Player(1)).iter().count(), 3);
-/// #   assert_eq!(query.at(&Player(2)).iter().count(), 4);
-/// }
-/// # world.run_system_cached(find_all_player_one_entities);
-/// ```
-pub struct QueryByIndex<'world, C: IndexableComponent, D: QueryData, F: QueryFilter = ()> {
-    world: UnsafeWorldCell<'world>,
-    state: Option<QueryState<D, (F, With<C>)>>,
-    last_run: Tick,
-    this_run: Tick,
-    index: Res<'world, Index<C>>,
-}
-
-impl<C: IndexableComponent, D: QueryData, F: QueryFilter> QueryByIndex<'_, C, D, F> {
-    /// Return a [`Query`] only returning entities with a component `C` of the provided value.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use bevy_ecs::prelude::*;
-    /// # let mut world = World::new();
-    /// #[derive(Component, PartialEq, Eq, Hash, Clone)]
-    /// #[component(immutable)]
-    /// enum FavoriteColor {
-    ///     Red,
-    ///     Green,
-    ///     Blue,
-    /// }
-    ///
-    /// world.add_index::<FavoriteColor>();
-    ///
-    /// fn find_red_fans(mut query: QueryByIndex<FavoriteColor, Entity>) {
-    ///     for entity in query.at(&FavoriteColor::Red).iter() {
-    ///         println!("{entity:?} likes the color Red!");
-    ///     }
-    /// }
-    /// ```
-    pub fn at(&mut self, value: &C) -> Query<'_, '_, D, (F, With<C>)> {
-        self.state = {
-            // SAFETY: Mutable references do not alias and will be dropped after this block
-            let mut builder = unsafe { QueryBuilder::new(self.world.world_mut()) };
-
-            match self.index.mapping.get(value) {
-                // If there is a marker, restrict to it
-                Some(state) => builder.with_id(state.component_id),
-                // Otherwise, create a no-op query by including With<C> and Without<C>
-                None => builder.without::<C>(),
-            };
-
-            Some(builder.build())
-        };
-
-        // SAFETY: We have registered all of the query's world accesses,
-        // so the caller ensures that `world` has permission to access any
-        // world data that the query needs.
-        unsafe {
-            Query::new(
-                self.world,
-                self.state.as_mut().unwrap(),
-                self.last_run,
-                self.this_run,
-            )
-        }
-    }
-}
-
-// SAFETY: We rely on the known-safe implementations of `SystemParam` for `Res` and `Query`.
-unsafe impl<C: IndexableComponent, D: QueryData + 'static, F: QueryFilter + 'static> SystemParam
-    for QueryByIndex<'_, C, D, F>
-{
-    type State = (QueryState<D, (F, With<C>)>, ComponentId);
-    type Item<'w, 's> = QueryByIndex<'w, C, D, F>;
-
-    fn init_state(world: &mut World, system_meta: &mut SystemMeta) -> Self::State {
-        let query_state = <Query<D, (F, With<C>)> as SystemParam>::init_state(world, system_meta);
-        let res_state = <Res<Index<C>> as SystemParam>::init_state(world, system_meta);
-
-        (query_state, res_state)
-    }
-
-    #[inline]
-    unsafe fn validate_param(
-        (query_state, res_state): &Self::State,
-        system_meta: &SystemMeta,
-        world: UnsafeWorldCell,
-    ) -> bool {
-        let query_valid = <Query<D, (F, With<C>)> as SystemParam>::validate_param(
-            query_state,
-            system_meta,
-            world,
-        );
-        let res_valid =
-            <Res<Index<C>> as SystemParam>::validate_param(res_state, system_meta, world);
-
-        query_valid && res_valid
-    }
-
-    unsafe fn get_param<'world, 'state>(
-        (query_state, res_state): &'state mut Self::State,
-        system_meta: &SystemMeta,
-        world: UnsafeWorldCell<'world>,
-        change_tick: Tick,
-    ) -> Self::Item<'world, 'state> {
-        query_state.validate_world(world.id());
-
-        let index =
-            <Res<Index<C>> as SystemParam>::get_param(res_state, system_meta, world, change_tick);
-
-        QueryByIndex {
-            world,
-            state: None,
-            last_run: system_meta.last_run,
-            this_run: change_tick,
-            index,
-        }
-    }
-}
 
 /// This [`Resource`] is responsible for managing a value-to-[`ComponentId`] mapping, allowing
 /// [`QueryByIndex`] to simply filter by [`ComponentId`] on a standard [`Query`].
@@ -465,4 +316,29 @@ impl<C: IndexableComponent> Index<C> {
 
         world.register_component_with_descriptor(descriptor)
     }
+
+    fn filter_query_for<D: QueryData, F: QueryFilter>(&self, builder: &mut QueryBuilder<D, F>, value: &C) {
+        let Some(state) = self.mapping.get(value) else {
+            // If there is no marker, create a no-op query by including With<C> and Without<C>
+            builder.without::<C>();
+            return;
+        };
+
+        // If there is a marker, restrict to it
+        builder.with_id(state.component_id);
+    }
+}
+
+/// Converts a [`usize`] into a [`bool`] array.
+const fn usize_to_bool_array(value: usize) -> [bool; size_of::<usize>() * 8] {
+    const LENGTH: usize = size_of::<usize>() * 8;
+    let mut array = [false; LENGTH];
+
+    let mut index = 0;
+    while index < LENGTH {
+        array[index] = value & (1 << index) > 0;
+        index += 1;
+    }
+
+    array
 }
